Refactored components/PlayMode
