+++ src/components/Game.js
@@ -3,10 +3,10 @@ import { Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import GameClock from '../features/gameClock/GameClock';
 import StartedButtonsGrandmasterMode from '../features/grandmasterMode/StartedButtonsGrandmasterMode';
-import FinishedButtonsPlayMode from './PlayMode/FinishedButtonsPlayMode';
-import FinishedDialogsPlayMode from './PlayMode/FinishedDialogsPlayMode';
-import StartedButtonsPlayMode from './PlayMode/StartedButtonsPlayMode';
-import StartedDialogsPlayMode from './PlayMode/StartedDialogsPlayMode';
+import FinishedButtonsPlayMode from '../features/playMode/FinishedButtonsPlayMode';
+import FinishedDialogsPlayMode from '../features/playMode/FinishedDialogsPlayMode';
+import StartedButtonsPlayMode from '../features/playMode/StartedButtonsPlayMode';
+import StartedDialogsPlayMode from '../features/playMode/StartedDialogsPlayMode';
 import MoveValidatorTable from '../features/table/MoveValidatorTable.js';
 import SecondaryButtons from './SecondaryButtons.js';
 import SecondaryDialogs from './SecondaryDialogs';
